+++ gdpr/src/main/java/com/kairos/response/dto/common/AssessmentBasicResponseDTO.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.kairos.dto.gdpr.Staff;
 import com.kairos.enums.gdpr.AssessmentSchedulingFrequency;
 import com.kairos.enums.gdpr.AssessmentStatus;
-import java.math.BigInteger;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -13,7 +13,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AssessmentBasicResponseDTO {
 
-    private BigInteger id;
+    private Long id;
     private String name;
     private LocalDate endDate;
     private LocalDate completedDate;
@@ -30,9 +30,9 @@ public class AssessmentBasicResponseDTO {
 
     public void setRisks(List<RiskBasicResponseDTO> risks) { this.risks = risks; }
 
-    public BigInteger getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(BigInteger id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getName() { return name; }
